Remove MetaDataServiceImpl dependencies from external modules
